requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-http-client==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
sendgrid==6.1.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
